Creates a sequential file with line feed
This program exemplifies the use of several GNUCOBOL 3.2 instructions such as using the screen to display and accept values, and creating a sequential file with a line feed delimiter for each record, regardless of whether such a file has been created previously, and simulates it as if it were a report. It is a first approach to advanced COBOL testing.