Program for processing indexed sequential files.
Preliminary prototype program that is responsible for maintaining a sequential file indexed with two fields and that has a menu for maintenance in the beginning. There is a problem when repositioning the sequential reading pointer of the file when a key does not exist and therefore there is an abnormal termination.