Program that counts characters in a string
This program has been optimized to have two independent tables, one with the predefined ASCII characters and another to load the characters to be searched for in a string, and thus display character count statistics. It was implemented this way so that in the future arbitrary characters can be loaded in another order to be searched for in a given string.