Employee Maintenance Program
This program provides a very complete maintenance of an employee file with two fields (code and salary), whose organization is sequentially indexed, its access is dynamic and its opening mode is input and output. It should be noted that the code field is the primary key and the salary field is an alternative key. It is a very versatile and powerful program so far.