Employee File Maintenance Program
This program aims to provide very complete maintenance for a simple employee file with only two fields (code and salary), and whose keys are, respectively, primary and alternate fields. The organization of this file is sequential and indexed, its access mode is dynamic and its opening mode is input and output.