Employee File Maintenance Program
This program provides complete and comprehensive maintenance for a simple employee file with two fields (code and salary), whose characteristics are to be a primary key and an alternative key, respectively. This employee file is indexed in a sequential organization, its access mode is dynamic and its opening mode is input and output.