Employee File Maintenance Program
The purpose of this program is to provide a comprehensive and complete maintenance of additions, deletions, changes and queries to a simple employee file with only two fields (code and salary), whose characteristics are, respectively, being a primary key and an alternative key each of them. This file has a sequential indexed organization, its access mode is dynamic and its opening mode is input and output.