Employee File Maintenance Program
This program provides very complete and extensive maintenance for a simple employee file with only two fields (code and salary) that are, respectively, primary key and alternative key. This employee file is indexed sequentially organized, its access mode is dynamic and its opening mode is input and output. It has very advanced error verification controls throughout the code of this program.