Given Date Validates Program
Warning! This program requires you to check the calculation that is performed to calculate the day of the week for a given date given its respective day, month, and year. The calculation appears to be out of date. The program itself validates whether a date is correct and returns the verified date as the result.