Date Validation & Get Day Of Week Program
The purpose of this program is to validate the data entered for a given date (day, month and year), which corresponds to the normal limits for each data (year >= 1582, month >=1 and <= 12, days >=1 and <= limit per month). Once these validation filters have been passed, the name of the day of the week that corresponds to the successfully validated date is calculated using the Perpetual Calendar.