Mathematical Functions General Project
The primary purpose of this program is to record a linear sequential output file with a variable length of up to 31 characters. Each record has a fixed first two positions containing the exact number of characters to be extracted. It also includes lineage control for generating reports and page feeds at set intervals. This program is based on filex2.cbl and is ideal for learning the GNU COBOL 3.2 language.

Manipulate programs sequential files.
These programs have been improved and optimized to work dynamically with sequential organization and indexed sequential organization files, in order to practice the GNUCOBOL 3.2 programming language.

Creates a sequential file with line feed.
Some minor fixes were made to classes and data file input-output definition.