Mathematical Functions General Project
Sample COBOL Screen Program.
The purpose of this program is to demonstrate the use of the COBOL screen section so that interesting screens can be displayed with variable positioning for a simple employee data entry. Each elementary screen data name has a data name to position the labels in varying rows and columns.

Text Windows Frames Program.
This program automatically generates text frames to simulate windows in the same text mode, changing only some input variables that are used in various program routines.